extract build meta info into utils, add error handling
